+++ includes/services/GoogleMaps3/jquery.googlemap.js
@@ -50,6 +50,23 @@
 		 */
 		this.imageoverlays = [];
 
+		/**
+		/* this is used to change the color of spiderable markers
+		*/
+		this.spiderMarkers = [];
+
+		/**
+		/* this global tracks the last infowindow opened so we can close it to spiderfy overlapping markers
+		*/
+		var lastInfoWindow = null;
+
+		/**
+		/* icon for spiderable markers
+		*/
+		_this.spiderMarkerIcon = mw.config.get('wgScriptPath')+'/extensions/Maps/includes/services/GoogleMaps3/img/yellow-dot.png';
+		_this.spiderNumberedIcon = mw.config.get('wgScriptPath')+'/extensions/Maps/includes/services/GoogleMaps3/img/num-markers/';
+
+
 
 		var getBounds = function() {
 			if (( options.centre === false || options.zoom === false ) && options.locations.length > 1) {
@@ -614,21 +631,76 @@
 			 * Allows grouping of markers.
 			 */
 			if ( options.markercluster ) {
+				_this.clusterOn = true;
 				mw.loader.using(
 					'ext.maps.gm3.markercluster',
 					function() {
 						_this.markercluster = new MarkerClusterer( _this.map, _this.markers, {
-							averageCenter: true
+							averageCenter: true,
+							maxZoom: options.markerclustermaxzoom,
+							ignoreHidden: true
 						} );
+					}
+				);
+
+				// enable OverlappingMarkerSpiderfier beyond markerclustermaxzoom
+				mw.loader.using(
+					'ext.maps.gm3.oms',
+					function() {
+						_this.oms = new OverlappingMarkerSpiderfier(_this.map, {
+							keepSpiderfied: true,
+							nearbyDistance: 30,
+							circleFootSeparation: 40,
+							spiralLengthFactor: 5,
+							spiralFootSeparation: 28
+						} );
+
+						for (var i = 0; i < _this.markers.length; i++) {
+							_this.oms.addMarker(_this.markers[i]);
+						}
+
+						//Add oms spiderfy listener to close the last infowindow when we spiderfy, and to show the count of spiderfied markers
+						_this.oms.addListener('spiderfy', function (markers) {
+							lastInfoWindow.close();
+							var i;
+							for	(i = 0; i < markers.length - 1; i++) {
+						    	markers[i].setIcon(_this.spiderNumberedIcon + 'yellow/marker' + ( i<98 ? (i+1) : '99plus') + '.png');
+							}
+							// make the marker holding the count of spiderfied markers green
+							markers[i].setIcon(_this.spiderNumberedIcon + 'green/marker' + ( i<98 ? (i+1) : '99plus') + '.png');
+						});
+
+						google.maps.event.addListener(_this.map, 'zoom_changed', function() {
+							// check if we're at the right zoom level where spidermarkers are visible
+							if (_this.map.getZoom()+1 < options.markerclustermaxzoom && _this.clusterOn) return;
+							setTimeout(function() {
+							    this.spiderMarkers = _this.oms.markersNearAnyOtherMarker();
+								for	(var i = 0; i < this.spiderMarkers.length; i++) {
+									var currIcon = this.spiderMarkers[i].getIcon();
+									if (typeof currIcon === 'string' && currIcon.indexOf(_this.spiderNumberedIcon) == 0) continue;
+								    this.spiderMarkers[i].setIcon(_this.spiderMarkerIcon);
+								}
+							}, 500);
+						});
+
+						google.maps.event.addListener(_this.map, 'dragend', function() {
+							// check if we're at the right zoom level where spidermarkers are visible
+							if (_this.map.getZoom()+1 < options.markerclustermaxzoom && _this.clusterOn) return;
+							setTimeout(function() {
+							    this.spiderMarkers = _this.oms.markersNearAnyOtherMarker();
+								for	(var i = 0; i < this.spiderMarkers.length; i++) {
+									var currIcon = this.spiderMarkers[i].getIcon();
+									if (typeof currIcon === 'string' && currIcon.indexOf(_this.spiderNumberedIcon) == 0) continue;
+								    this.spiderMarkers[i].setIcon(_this.spiderMarkerIcon);								}
+							}, 500);
+						});
 					}
 				);
 			}
 
-
-
 			if (options.searchmarkers) {
 				var searchBoxValue = mediaWiki.msg('maps-searchmarkers-text');
-				var searchBox = $('<input type="text" value="' + searchBoxValue + '" />');
+				var searchBox = $('<input id="searchBoxGMap" type="text" value="' + searchBoxValue + '" />');
 				var searchContainer = document.createElement('div');
 				searchContainer.style.padding = '5px';
 				searchContainer.index = 1;
@@ -640,8 +712,53 @@
 					map.panBy(0,-30);
 				});
 
+				if (options.markercluster) {
+					var clusterToggle = $('<label for="clusterToggle">Cluster</label><input id="clusterToggle" type="checkbox"/>');
+					var clusterContainer = document.createElement('div');
+					clusterContainer.style.padding = '5px';
+					clusterToggle.appendTo(clusterContainer);
+					clusterToggle.prop('checked', true);
+					map.controls[google.maps.ControlPosition.TOP_RIGHT].push(clusterContainer);
+
+					clusterToggle.on('click', function() {
+					    var $this = $(this); 
+					    if ($this.is(':checked')) {
+					        // recluster 
+							for (var i = 0; i < _this.markers.length; i++) { 
+								_this.markers[i].setIcon('');
+					        }
+					        _this.markercluster = new MarkerClusterer( _this.map, _this.markers, {
+								averageCenter: false,
+								maxZoom: options.markerclustermaxzoom,
+								ignoreHidden: true
+							} );
+							_this.clusterOn = true;
+					    } else {
+					    	// uncluster
+					        _this.markercluster.clearMarkers();
+							for (var i = 0; i < _this.markers.length; i++) { 
+								_this.markers[i].setOptions({map: _this.map, visible:true});
+					        }
+					        _this.clusterOn = false;
+					        document.getElementById('searchBoxGMap').value = '';
+					        google.maps.event.trigger(map, 'zoom_changed');
+					    }
+					});
+				}
+
 
 				searchBox.on('keyup',function (e) {
+
+					// automatically turn off markercluster when searching markers
+			        if (options.markercluster) {
+			        	_this.markercluster.clearMarkers();
+						for (var i = 0; i < _this.markers.length; i++) { 
+							_this.markers[i].setOptions({map: _this.map, visible:true});
+				        }
+				        clusterToggle.prop('checked', false);
+				        google.maps.event.trigger(map, 'zoom_changed');
+			        }
+
 					for (var i = 0; i < _this.markers.length; i++) {
 						var haystack = '';
 						var marker = _this.markers[i];
@@ -845,6 +962,7 @@
 			}
 			else {
 				this.openWindow.open( this.map );
+				lastInfoWindow = this.openWindow;
 			}
 		}